requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.0.4
attrs==18.2.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
docopt==0.6.2
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysql-connector==2.1.6
odfpy==1.3.6
openpyxl==2.5.8
pkg-resources==0.0.0
requests==2.20.1
six==1.11.0
tablib==0.12.1
toml==0.10.0
typed-ast==1.1.0
unicodecsv==0.14.1
urllib3==1.24.1
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
